# Use unpacking inside arrays