--- a/src/main/php/PDepend/Source/AST/ASTClassOrInterfaceReference.php
+++ b/src/main/php/PDepend/Source/AST/ASTClassOrInterfaceReference.php
@@ -86,7 +86,7 @@ class ASTClassOrInterfaceReference extends ASTType
      */
     public function __sleep()
     {
-        return array_merge(['context'], parent::__sleep());
+        return ['context', ...parent::__sleep()];
     }
 
     /**