--- a/src/main/php/PDepend/Source/AST/ASTConstantDeclarator.php
+++ b/src/main/php/PDepend/Source/AST/ASTConstantDeclarator.php
@@ -93,7 +93,7 @@ class ASTConstantDeclarator extends AbstractASTNode
      */
     public function __sleep()
     {
-        return array_merge(['value'], parent::__sleep());
+        return ['value', ...parent::__sleep()];
     }
 
     /**