--- a/src/main/php/PDepend/Source/AST/ASTIncludeExpression.php
+++ b/src/main/php/PDepend/Source/AST/ASTIncludeExpression.php
@@ -69,7 +69,7 @@ class ASTIncludeExpression extends ASTExpression
      */
     public function __sleep()
     {
-        return array_merge(['once'], parent::__sleep());
+        return ['once', ...parent::__sleep()];
     }
 
     /**