--- a/src/main/php/PDepend/Source/AST/ASTParentReference.php
+++ b/src/main/php/PDepend/Source/AST/ASTParentReference.php
@@ -80,7 +80,7 @@ final class ASTParentReference extends ASTClassOrInterfaceReference
      */
     public function __sleep()
     {
-        return array_merge(['reference'], parent::__sleep());
+        return ['reference', ...parent::__sleep()];
     }
 
     /**