--- a/src/main/php/PDepend/Source/AST/ASTRequireExpression.php
+++ b/src/main/php/PDepend/Source/AST/ASTRequireExpression.php
@@ -69,7 +69,7 @@ class ASTRequireExpression extends ASTExpression
      */
     public function __sleep()
     {
-        return array_merge(['once'], parent::__sleep());
+        return ['once', ...parent::__sleep()];
     }
 
     /**