--- a/src/main/php/PDepend/Source/AST/ASTVariableDeclarator.php
+++ b/src/main/php/PDepend/Source/AST/ASTVariableDeclarator.php
@@ -67,7 +67,7 @@ class ASTVariableDeclarator extends ASTExpression
      */
     public function __sleep()
     {
-        return array_merge(['value'], parent::__sleep());
+        return ['value', ...parent::__sleep()];
     }
 
     /**