--- a/src/main/php/PDepend/Source/Language/PHP/AbstractPHPParser.php
+++ b/src/main/php/PDepend/Source/Language/PHP/AbstractPHPParser.php
@@ -625,12 +625,12 @@ abstract class AbstractPHPParser
 
             // Add mapping between image and qualified name to symbol table
             if ($image !== false) {
-                $this->useSymbolTable->add($image, implode('', array_merge($fragments, $subFragments)));
+                $this->useSymbolTable->add($image, implode('', [...$fragments, ...$subFragments]));
             }
         }
 
         if (isset($image, $subFragments) && $image !== false) {
-            $this->useSymbolTable->add($image, implode('', array_merge($fragments, $subFragments)));
+            $this->useSymbolTable->add($image, implode('', [...$fragments, ...$subFragments]));
         }
 
         $this->consumeToken(Tokens::T_CURLY_BRACE_CLOSE);