Type: refactoring
Breaking change: no

With unpacking in arrays (introduced in PHP 7.4) `array_merge()` can now be replaced by operators.