@ravage84 2024-05-12
Hm... what is the actual goal of this one? Or what is the problem with `array_merge`?

Have you considered the varying behavior of both ways?

https://stitcher.io/blog/what-is-array-plus-in-php

https://www.php.net/manual/en/function.array-merge.php
https://www.php.net/manual/en/language.operators.array.php


Also, if [this comment](https://www.php.net/manual/en/function.array-merge.php#126687) is true, it might even be a minor performance downgrade.

🤔