@AJenbo 2024-05-12
> Hm... what is the actual goal of this one? Or what is the problem with array_merge?

I find the new syntax cleaner. Same with array union. Sadly none of them could be reduced to `+=` but still:
```php
$items = $proportions + $metrics;
```
Then the `array_merge()` variant:
```php
$items = array_merge($metrics, $proportions);`
```

For the splat operators I particularly find this one a big improvment:
```php
return ['metadata', ...parent::__sleep()];
```
over `array_merge()` since it doesn't create an array just to satisfy an interface:
```php
return array_merge(['metadata'], parent::__sleep());
```

In PHPMD you will also see some where it is nested in other calls which can make it a bit harder to see what is happening when using array_merge:

```php
$args = array_unique(array_filter(array_merge([__CLASS__, $this::class], $this->getClasses())));
```

Where the unpacking makes it more clear that is where since the square brackets wrap the array bits:
```php
$args = array_unique(array_filter([__CLASS__, $this::class, ...$this->getClasses()]));
```


I also like this one  as i don't have to look at the end of the first line to understand that we are merging, that becomes clear from each of the lines that are being merged:

```php
->will(static::returnValue([
	...array_map([$this, 'createPublicMethod'], $publicMethods),
	...array_map([$this, 'createPrivateMethod'], $privateMethods),
]));
```
vs:
```php
->will(static::returnValue(array_merge(
	array_map([$this, 'createPublicMethod'], $publicMethods),
	array_map([$this, 'createPrivateMethod'], $privateMethods)
)));
```

So to sum it up, for me it's similar to the use of `[]` instead of `array()`. There is no big impact of it it's just new syntax to make things a little bit nicer.

In most cases the lines are shorter, there are few lines, or over all fewer symbols. It's only when you are merging 4 arrays that you would start it would start to get longer then `array_merge()`  and we don't have any such cases.

> Have you considered the varying behavior of both ways?

Yes

> Also, if [this comment](https://www.php.net/manual/en/function.array-merge.php#126687) is true, it might even be a minor performance downgrade.

I ran a few yest and performance was identical.

Running the one from the comment I actually get better performance using spread then `array_merge()`:

```
0.38876485824585 sec for spread
0.43788719177246 sec for array_merge
```

I also added a test for the inline case where the performance is even better:

```php
$before = microtime(true);

for ($i=0 ; $i<10000000 ; $i++) {
    $array2 = ['carrot','lettuce','broccoli'];

    $array1 = ['apple','orange','banana',...$array2];
}

$after = microtime(true);
echo ($after-$before) . " sec for inline spread\n";
```

```
0.36439108848572 sec for inline spread
```
(I'm running PHP 8.3.6 here)