@AJenbo 2024-05-12
For completeness:
```
0.39206385612488 sec for spread
0.44558620452881 sec for array_merge
0.36626315116882 sec for inline spread
0.56836605072021 sec for array_push
0.57034206390381 sec for array_unshift
0.60777497291565 sec for foreach
```

I guess we can expect a 11-17% performance increase from this :shrug:
We should also hunt down any place where foreach is being used as it is very slow for combining arrays.

```php
<?php
$before = microtime(true);
for ($i=0 ; $i<10000000 ; $i++) {
    $array1 = ['apple','orange','banana'];
    $array2 = ['carrot','lettuce','broccoli'];

    $array1 = [...$array1,...$array2];
}
$after = microtime(true);
echo ($after-$before) . " sec for spread\n";

$before = microtime(true);
for ($i=0 ; $i<10000000 ; $i++) {
    $array1 = ['apple','orange','banana'];
    $array2 = ['carrot','lettuce','broccoli'];

    $array1 = array_merge($array1,$array2);
}
$after = microtime(true);
echo ($after-$before) . " sec for array_merge\n";

$before = microtime(true);
for ($i=0 ; $i<10000000 ; $i++) {
    $array2 = ['carrot','lettuce','broccoli'];

    $array1 = ['apple','orange','banana',...$array2];
}
$after = microtime(true);
echo ($after-$before) . " sec for inline spread\n";

$before = microtime(true);
for ($i=0 ; $i<10000000 ; $i++) {
    $array1 = ['apple','orange','banana'];
    $array2 = ['carrot','lettuce','broccoli'];

    array_push($array1,...$array2);
}
$after = microtime(true);
echo ($after-$before) . " sec for array_push\n";

$before = microtime(true);
for ($i=0 ; $i<10000000 ; $i++) {
    $array1 = ['apple','orange','banana'];
    $array2 = ['carrot','lettuce','broccoli'];

    array_unshift($array2,$array1);
}
$after = microtime(true);
echo ($after-$before) . " sec for array_unshift\n";

$before = microtime(true);
for ($i=0 ; $i<10000000 ; $i++) {
    $array1 = ['apple','orange','banana'];
    $array2 = ['carrot','lettuce','broccoli'];
	foreach ($array2 as $value) {
		$array1[] = $value;
	}
}
$after = microtime(true);
echo ($after-$before) . " sec for foreach\n";
```