@AJenbo 2024-05-12
Older versions via 3v4l.org (and a lower sample size to allow it to finish in the given time limit)

PHP 8.2.19
```php
0.0014340877532959 sec for spread
0.0015029907226562 sec for array_merge
0.0013399124145508 sec for inline spread
0.0018529891967773 sec for array_push
0.0018692016601562 sec for array_unshift
0.0018501281738281 sec for foreach
```

PHP 8.1.28:
```php
0.00092291831970215 sec for spread
0.0010159015655518 sec for array_merge
0.00088095664978027 sec for inline spread
0.0012168884277344 sec for array_push
0.0013630390167236 sec for array_unshift
0.0012340545654297 sec for foreach
```