@ravage84 2024-05-12
While I'm sold on the splat operator, I'm not so much on the plus operator (union operation) on arrays.

Reading ``array_merge`` makes the intent clear and it's easier to reach the PHP doc on that one, in case one needs to re-read the behavior specifics versus the ones of some plus sign on arrays.

Also, applying newer, less character containing syntax to make convoluted code more readable is more of a workaround than the way to go for me.

Lastly, I'm not sure we introduce minor changes in behavior in the name of getting a bit more readable code since union and merge isn't the same after all because we might not have some edge cases not covered by tests.

Anyway, I'm eager to hear what the others think about this.

Thank you for the performance tests.