@AJenbo 2024-05-12
One thing I prefer from using union over array_merge is that the intent to keep elements or overwrite them becomes explicit. With `array_merge()` it might combine, overwrite, or even a combination of the two depending on the type of keys it detects.